+++ src/cube_factory.cpp
@@ -110,10 +110,10 @@ void cube_factory::register_default() {
     cube_generators.insert(std::make_pair<std::string, std::function<std::shared_ptr<cube>(nlohmann::json&)>>(
         "apply_pixel", [](nlohmann::json& j) {
             if (j.count("band_names") > 0) {
-                auto x = apply_pixel_cube::create(instance()->create_from_json(j["in_cube"]), j["expr"].get<std::vector<std::string>>(), j["band_names"].get<std::vector<std::string>>());
+                auto x = apply_pixel_cube::create(instance()->create_from_json(j["in_cube"]), j["expr"].get<std::vector<std::string>>(), j["band_names"].get<std::vector<std::string>>(), j["keep_bands"].get<bool>());
                 return x;
             } else {
-                auto x = apply_pixel_cube::create(instance()->create_from_json(j["in_cube"]), j["expr"].get<std::vector<std::string>>());
+                auto x = apply_pixel_cube::create(instance()->create_from_json(j["in_cube"]), j["expr"].get<std::vector<std::string>>(), {}, j["keep_bands"].get<bool>());
                 return x;
             }
         }));